FABN-1598 Check discovery endorsements (#314)
Have the discovery handler check the endorsement responses
to see if the read/write sets are all the same, try another
layout when not the same.
